+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # Changelog
 
 ## [Unreleased]
-### No Changes
+### Fixed
+- Contact Confirmation email
+- Bug with pageConfig injection.
 
 ## [4.0.0-RC1] - 2019-07-14
 ## Notes